OutputStandardScale(US#35)
Implemented DurationToHeightConverterTest, GapToMarginConverterTest, SequenceDiagram, DurationToHeightConverter, GapToMarginConverter.

Now time scale is 5 seconds per tab height.

+++ src/GraphalyzerPro/GraphalyzerPro.SequenceDiagramAnalysis.Tests/GapToMarginConverterTest.cs
@@ -49,28 +49,28 @@ namespace GraphalyzerPro.SequenceDiagramAnalysis.Tests
         public void Convert_AllValuesOk()
         {
             var converter = new GapToMarginConverter();
-            ((Thickness)(converter.Convert(new object[] { (long)(1), (long)(10), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(2.0);
+            ((Thickness)(converter.Convert(new object[] { (long)(1), (double)(10), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(2.0);
         }
 
         [Test]
         public void Convert_AllValuesOk_ParameterNull()
         {
             var converter = new GapToMarginConverter();
-            ((Thickness)(converter.Convert(new object[] { (long)(1), (long)(10), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Left.Should().Be(0.0);
+            ((Thickness)(converter.Convert(new object[] { (long)(1), (double)(10), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Left.Should().Be(0.0);
         }
 
         [Test]
         public void Convert_AllValuesOk_ParameterSet()
         {
             var converter = new GapToMarginConverter();
-            ((Thickness)(converter.Convert(new object[] { (long)(1), (long)(10), (double)(20) }, typeof(double), (long)(5), CultureInfo.InvariantCulture))).Left.Should().Be(5.0);
+            ((Thickness)(converter.Convert(new object[] { (long)(1), (double)(10), (double)(20) }, typeof(double), (long)(5), CultureInfo.InvariantCulture))).Left.Should().Be(5.0);
         }
 
         [Test]
         public void Convert_FirstValueDependencyPropertyUnset()
         {
             var converter = new GapToMarginConverter();
-            ((Thickness)(converter.Convert(new object[] { DependencyProperty.UnsetValue, (long)(10), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(0.0);
+            ((Thickness)(converter.Convert(new object[] { DependencyProperty.UnsetValue, (double)(10), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(0.0);
         }
 
         [Test]
@@ -84,14 +84,14 @@ namespace GraphalyzerPro.SequenceDiagramAnalysis.Tests
         public void Convert_ThirdValueDependencyPropertyUnset()
         {
             var converter = new GapToMarginConverter();
-            ((Thickness)(converter.Convert(new object[] { (long)(1), (long)(10), DependencyProperty.UnsetValue }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(0.0);
+            ((Thickness)(converter.Convert(new object[] { (long)(1), (double)(10), DependencyProperty.UnsetValue }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(0.0);
         }
 
         [Test]
         public void Convert_TotalDurationIsNull()
         {
             var converter = new GapToMarginConverter();
-            ((Thickness)(converter.Convert(new object[] { (long)(1), (long)(0), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(0.0);
+            ((Thickness)(converter.Convert(new object[] { (long)(1), (double)(0), (double)(20) }, typeof(double), null, CultureInfo.InvariantCulture))).Top.Should().Be(0.0);
         }
     }
 }
